внесены синтаксические правки
