+++ js/form/on-edit-modal-control.js
@@ -1,10 +1,12 @@
 import {isEscapeKey} from '../util.js';
+import { Slider } from './slider.js';
 
 const form = document.querySelector('.img-upload__form');
 const inputUploadFile = form.querySelector('#upload-file');
 const imgOverlay = form.querySelector('.img-upload__overlay');
 const uploadCancel = form.querySelector('#upload-cancel');
 const imgPreview = document.querySelector('.img-upload__preview img');
+const blockImgPreview = document.querySelector('.img-upload__preview');
 
 const onEditModalControl = () => {
 
@@ -25,6 +27,8 @@ const onEditModalControl = () => {
     document.body.classList.remove('modal-open');
     form.reset();
     imgPreview.className = '';
+    blockImgPreview.style.transform = `scale(${1})`;
+    Slider.destroy();
   }
 
   inputUploadFile.addEventListener('input', () => {
